# comment/api/comment/v1/comment_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.8.4
// - protoc             v5.28.3
// source: api/comment/v1/comment.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationCommentCreateComment = "/comment.v1.Comment/CreateComment"
const OperationCommentDeleteComment = "/comment.v1.Comment/DeleteComment"
const OperationCommentGetCommentList = "/comment.v1.Comment/GetCommentList"
const OperationCommentGetCommentListWithReplies = "/comment.v1.Comment/GetCommentListWithReplies"
const OperationCommentGetReplyList = "/comment.v1.Comment/GetReplyList"
const OperationCommentLikeComment = "/comment.v1.Comment/LikeComment"

type CommentHTTPServer interface {
	// CreateComment CreateComment 发表评论
	CreateComment(context.Context, *CreateCommentRequest) (*CreateCommentReply, error)
	// DeleteComment DeleteComment 删除评论
	DeleteComment(context.Context, *DeleteCommentRequest) (*DeleteCommentReply, error)
	// GetCommentList GetCommentList 获取一级评论列表
	GetCommentList(context.Context, *GetCommentListRequest) (*GetCommentListReply, error)
	// GetCommentListWithReplies GetCommentListWithReplies 获取一级评论列表及其前两条二级评论
	GetCommentListWithReplies(context.Context, *GetCommentListWithRepliesRequest) (*GetCommentListWithRepliesReply, error)
	// GetReplyList GetReplyList 获取二级评论列表
	GetReplyList(context.Context, *GetReplyListRequest) (*GetReplyListReply, error)
	// LikeComment LikeComment 点赞评论
	LikeComment(context.Context, *LikeCommentRequest) (*LikeCommentReply, error)
}

func RegisterCommentHTTPServer(s *http.Server, srv CommentHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/comment", _Comment_CreateComment0_HTTP_Handler(srv))
	r.DELETE("/v1/comment", _Comment_DeleteComment0_HTTP_Handler(srv))
	r.POST("/v1/comment/like", _Comment_LikeComment0_HTTP_Handler(srv))
	r.GET("/v1/comment/list", _Comment_GetCommentList0_HTTP_Handler(srv))
	r.GET("/v1/comment/list/replies", _Comment_GetCommentListWithReplies0_HTTP_Handler(srv))
	r.GET("/v1/comment/replies", _Comment_GetReplyList0_HTTP_Handler(srv))
}

func _Comment_CreateComment0_HTTP_Handler(srv CommentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateCommentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCommentCreateComment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateComment(ctx, req.(*CreateCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateCommentReply)
		return ctx.Result(200, reply)
	}
}

func _Comment_DeleteComment0_HTTP_Handler(srv CommentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteCommentRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCommentDeleteComment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteComment(ctx, req.(*DeleteCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteCommentReply)
		return ctx.Result(200, reply)
	}
}

func _Comment_LikeComment0_HTTP_Handler(srv CommentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LikeCommentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCommentLikeComment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.LikeComment(ctx, req.(*LikeCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LikeCommentReply)
		return ctx.Result(200, reply)
	}
}

func _Comment_GetCommentList0_HTTP_Handler(srv CommentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCommentListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCommentGetCommentList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCommentList(ctx, req.(*GetCommentListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetCommentListReply)
		return ctx.Result(200, reply)
	}
}

func _Comment_GetCommentListWithReplies0_HTTP_Handler(srv CommentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCommentListWithRepliesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCommentGetCommentListWithReplies)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCommentListWithReplies(ctx, req.(*GetCommentListWithRepliesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetCommentListWithRepliesReply)
		return ctx.Result(200, reply)
	}
}

func _Comment_GetReplyList0_HTTP_Handler(srv CommentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetReplyListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCommentGetReplyList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetReplyList(ctx, req.(*GetReplyListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetReplyListReply)
		return ctx.Result(200, reply)
	}
}

type CommentHTTPClient interface {
	CreateComment(ctx context.Context, req *CreateCommentRequest, opts ...http.CallOption) (rsp *CreateCommentReply, err error)
	DeleteComment(ctx context.Context, req *DeleteCommentRequest, opts ...http.CallOption) (rsp *DeleteCommentReply, err error)
	GetCommentList(ctx context.Context, req *GetCommentListRequest, opts ...http.CallOption) (rsp *GetCommentListReply, err error)
	GetCommentListWithReplies(ctx context.Context, req *GetCommentListWithRepliesRequest, opts ...http.CallOption) (rsp *GetCommentListWithRepliesReply, err error)
	GetReplyList(ctx context.Context, req *GetReplyListRequest, opts ...http.CallOption) (rsp *GetReplyListReply, err error)
	LikeComment(ctx context.Context, req *LikeCommentRequest, opts ...http.CallOption) (rsp *LikeCommentReply, err error)
}

type CommentHTTPClientImpl struct {
	cc *http.Client
}

func NewCommentHTTPClient(client *http.Client) CommentHTTPClient {
	return &CommentHTTPClientImpl{client}
}

func (c *CommentHTTPClientImpl) CreateComment(ctx context.Context, in *CreateCommentRequest, opts ...http.CallOption) (*CreateCommentReply, error) {
	var out CreateCommentReply
	pattern := "/v1/comment"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationCommentCreateComment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *CommentHTTPClientImpl) DeleteComment(ctx context.Context, in *DeleteCommentRequest, opts ...http.CallOption) (*DeleteCommentReply, error) {
	var out DeleteCommentReply
	pattern := "/v1/comment"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCommentDeleteComment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *CommentHTTPClientImpl) GetCommentList(ctx context.Context, in *GetCommentListRequest, opts ...http.CallOption) (*GetCommentListReply, error) {
	var out GetCommentListReply
	pattern := "/v1/comment/list"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCommentGetCommentList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *CommentHTTPClientImpl) GetCommentListWithReplies(ctx context.Context, in *GetCommentListWithRepliesRequest, opts ...http.CallOption) (*GetCommentListWithRepliesReply, error) {
	var out GetCommentListWithRepliesReply
	pattern := "/v1/comment/list/replies"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCommentGetCommentListWithReplies))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *CommentHTTPClientImpl) GetReplyList(ctx context.Context, in *GetReplyListRequest, opts ...http.CallOption) (*GetReplyListReply, error) {
	var out GetReplyListReply
	pattern := "/v1/comment/replies"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCommentGetReplyList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *CommentHTTPClientImpl) LikeComment(ctx context.Context, in *LikeCommentRequest, opts ...http.CallOption) (*LikeCommentReply, error) {
	var out LikeCommentReply
	pattern := "/v1/comment/like"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationCommentLikeComment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}